refactor/Changing how the calculation is performed for better performance and creating unit and integration tests

+++ src/main/java/com/eduardo/alticci/services/AlticciService.java
@@ -1,6 +1,8 @@
 package com.eduardo.alticci.services;
 
+import java.math.BigInteger;
+
 public interface AlticciService {
 
-    Integer calculateAlticciValue(Integer num);
+    BigInteger calculateAlticciValue(Integer num);
 }
